--- src/receiver/middleware/SSLCheck.ts
@@ -1,12 +1,0 @@
-import { RequestHandler } from 'express';
-
-export default function (): RequestHandler {
-  return (req, res, next) => {
-    if (req.body && req.body.ssl_check) {
-      res.send();
-      return;
-    }
-
-    next();
-  };
-}
