+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from './actions';
 export * from './command';
 export * from './events';
 export * from './options';
+export * from './receiver';
